Proposed specification for Date#to_time
The behavior in this specification is a proposal for a future releases
Ruby 1.9 and 2.0.

Currently (1.9.3-p429, 2.0.0-p195), Date#to_time uses the civic month, day
and year to construct a Time object, even if the Date is using the Julian
calendar. In the case of a Julian Date, the result is a Time object with
a value (epoch seconds) which does not correspond to to the Julian date
from which it was derived. The current behavior is not symmetric with
Time#to_date, which always constructs the date corresponding to the Time's
year, month and day in the Gregorian or proleptic Gregorian calendar. As
a result:

  Date.new(1582, 10, 4).to_time.to_date != Date.new(1582, 10, 4)